requirements.txt
pytesseract
pdf2image
Pillow
openai
python-dotenv
rapidfuzz
langchain
tiktoken
